added SlotValueEquals, drafted RegexpGroupSlot, formatting a GroupSlot with an f-string, moved derivative slots into a different file
