+++ chatsky/__rebuild_pydantic_models__.py
@@ -1,10 +1,10 @@
 # flake8: noqa: F401
 
 from chatsky.core.service.types import ExtraHandlerRuntimeInfo, ComponentExecutionState
-from chatsky.core import Context, Script
+from chatsky.core import Context, Message, Script
 from chatsky.core.script import Node
 from chatsky.core.pipeline import Pipeline
-from chatsky.slots.slots import SlotManager
+from chatsky.slots.base_slots import SlotManager, FunctionSlot
 from chatsky.core.context import FrameworkData, ServiceState
 from chatsky.core.service import PipelineComponent
 
@@ -15,3 +15,4 @@ Context.model_rebuild()
 ExtraHandlerRuntimeInfo.model_rebuild()
 FrameworkData.model_rebuild()
 ServiceState.model_rebuild()
+FunctionSlot.model_rebuild()
