+++ chatsky/slots/__init__.py
@@ -1,1 +1,1 @@
-from chatsky.slots.slots import GroupSlot, ValueSlot, RegexpSlot, FunctionSlot
+from chatsky.slots.base_slots import GroupSlot, ValueSlot, RegexpSlot, FunctionSlot
