+++ tests/slots/test_slot_functions.py
@@ -6,10 +6,10 @@ import pytest
 from chatsky import Context
 from chatsky.core import BaseResponse, Node
 from chatsky.core.message import MessageInitTypes, Message
-from chatsky.slots.slots import ValueSlot, SlotNotExtracted, GroupSlot, SlotManager
+from chatsky.slots.base_slots import ValueSlot, SlotNotExtracted, GroupSlot, SlotManager
 from chatsky import conditions as cnd, responses as rsp, processing as proc
 from chatsky.processing.slots import logger as proc_logger
-from chatsky.slots.slots import logger as slot_logger
+from chatsky.slots.base_slots import logger as slot_logger
 from chatsky.responses.slots import logger as rsp_logger
