+++ tests/slots/test_slot_manager.py
@@ -1,6 +1,6 @@
 import pytest
 
-from chatsky.slots.slots import (
+from chatsky.slots.base_slots import (
     SlotManager,
     RegexpSlot,
     GroupSlot,
